Feat: input 컴마 alret 수정
